* refactored into package-by-feature * gradle dependency cleanup
