+++ service/build.gradle
@@ -26,27 +26,32 @@ repositories {
 }
 
 dependencies {
+	// Spring
 	implementation 'org.springframework.boot:spring-boot-starter-actuator'
 	implementation 'org.springframework.boot:spring-boot-starter-web'
 	implementation 'org.springframework.boot:spring-boot-starter-data-mongodb'
+
+	// Logging
 	implementation 'ch.qos.logback:logback-classic:1.2.3'
 	implementation 'ch.qos.logback.contrib:logback-json-classic:0.1.5'
 	implementation 'ch.qos.logback.contrib:logback-jackson:0.1.5'
-	implementation 'org.springframework.boot:spring-boot-starter-web'
+
+	// Cloud
 	implementation 'com.amazonaws:aws-java-sdk-s3:1.11.907'
 	implementation 'com.azure:azure-storage-blob:12.9.0'
 	implementation 'org.springframework:spring-jms:5.2.6.RELEASE'
 	implementation 'org.apache.qpid:qpid-jms-client:0.51.0'
+
+	// Utils
 	implementation 'org.apache.commons:commons-lang3:3.11'
-	testImplementation 'org.apache.commons:commons-lang3:3.11'
+
+	// Test
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testImplementation "io.rest-assured:rest-assured:4.2.0"
 	testImplementation "io.rest-assured:json-path:4.2.0"
 	testImplementation "io.rest-assured:xml-path:4.2.0"
 	testImplementation "org.assertj:assertj-core:3.18.1"
 	testImplementation "org.testcontainers:testcontainers:1.15.0"
-	testImplementation 'org.springframework:spring-jms:5.2.6.RELEASE'
-	testImplementation 'org.apache.qpid:qpid-jms-client:0.51.0'
 }
 
 test {
