+++ service/src/intTest/java/uk/nhs/adaptors/gp2gp/common/storage/StorageConnectorTest.java
@@ -1,11 +1,10 @@
-package uk.nhs.adaptors.gp2gp.common.storageconnectors;
+package uk.nhs.adaptors.gp2gp.common.storage;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-
 import uk.nhs.adaptors.gp2gp.extension.ActiveMQExtension;
 import uk.nhs.adaptors.gp2gp.extension.MongoDBExtension;
