+++ service/src/intTest/java/uk/nhs/adaptors/gp2gp/extension/MongoDBExtension.java
@@ -1,14 +1,13 @@
 package uk.nhs.adaptors.gp2gp.extension;
 
 import lombok.extern.slf4j.Slf4j;
-import uk.nhs.adaptors.gp2gp.container.MongoDbContainer;
-import uk.nhs.adaptors.gp2gp.repositorys.EhrExtractStatusRepository;
-
 import org.junit.jupiter.api.extension.BeforeAllCallback;
 import org.junit.jupiter.api.extension.BeforeEachCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.springframework.context.ApplicationContext;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
+import uk.nhs.adaptors.gp2gp.container.MongoDbContainer;
+import uk.nhs.adaptors.gp2gp.ehr.EhrExtractStatusRepository;
 
 @Slf4j
 public class MongoDBExtension implements BeforeAllCallback, BeforeEachCallback {
