+++ service/src/main/java/uk/nhs/adaptors/gp2gp/common/storage/S3StorageConnector.java
@@ -1,11 +1,11 @@
-package uk.nhs.adaptors.gp2gp.common.storageconnectors;
-
-import java.io.InputStream;
+package uk.nhs.adaptors.gp2gp.common.storage;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.S3Object;
+
+import java.io.InputStream;
 
 public class S3StorageConnector implements StorageConnector {
     private final AmazonS3 s3client;
