+++ service/src/main/java/uk/nhs/adaptors/gp2gp/common/storage/StorageConnectorException.java
@@ -1,4 +1,4 @@
-package uk.nhs.adaptors.gp2gp.common.storageconnectors;
+package uk.nhs.adaptors.gp2gp.common.storage;
 
 public class StorageConnectorException extends RuntimeException {
     public StorageConnectorException(String errorMessage, Throwable err) {
