+++ service/src/main/java/uk/nhs/adaptors/gp2gp/ehr/EhrExtractStatusRepository.java
@@ -1,11 +1,9 @@
-package uk.nhs.adaptors.gp2gp.repositorys;
-
-import java.util.Optional;
-
-import uk.nhs.adaptors.gp2gp.models.EhrExtractStatus;
+package uk.nhs.adaptors.gp2gp.ehr;
 
 import org.springframework.data.mongodb.repository.Query;
 import org.springframework.data.repository.CrudRepository;
+
+import java.util.Optional;
 
 public interface EhrExtractStatusRepository extends CrudRepository<EhrExtractStatus, String> {
     @Query("{ 'extractId' : ?0 }")
