+++ service/src/main/java/uk/nhs/adaptors/gp2gp/mhs/InboundMessage.java
@@ -1,10 +1,10 @@
-package uk.nhs.adaptors.gp2gp.models;
+package uk.nhs.adaptors.gp2gp.mhs;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
-public class MhsInboundMessage {
+public class InboundMessage {
     private String payload;
 }
